requirements.txt
langchain
pandas
langchain-experimental
openai
python-dotenv
tabulate
matplotlib
django
langchain-openai
langchain-ollama
